pagent-os/old-lib-src/requirements.txt
# Costco Scraper Requirements
# Install with: pip install -r requirements.txt

# Web scraping - Playwright-focused
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
fake-useragent>=1.4.0

# AI-powered neuromorphic ETL
google-generativeai>=0.8.0

# Environment configuration
python-dotenv>=1.0.0

# Data handling
pandas>=2.1.0
lxml>=4.9.0

# Flask API and documentation
flask>=2.3.0
flask-restx>=1.3.0
werkzeug>=2.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
pathlib2>=2.3.7; python_version < '3.4'

# Optional: For better HTML parsing
html5lib>=1.1
